bugfix #1: src page is wrong, correct ID is 42
Upstream MSDN page has wrong ID for PROGRAMFILESX86. I didn't look to
see if others are correct. It turns out ctypes is really much better --
when you have the right code to work with. That's coming in next commits.
So no point in carrying on with what I've been doing (no regrets about
the learning though. It's been rewarding.)